requirements.txt
Flask
scikit-learn
pandas
numpy
joblib
python-whois
